sdk/python/dagger/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "dagger"
version = "0.1.0"
authors = [
  { name="Sam Alba", email="sam@dagger.io" },
]
description = "Dagger library"
keywords = [
  "dagger",
]
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
classifiers = [
  "Development Status :: 3 - Alpha",
  "Framework :: Hatch",
  # "Framework :: AnyIO",
  "Framework :: aiohttp",
  "Framework :: Pytest",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.10",
  "License :: OSI Approved :: Apache Software License",
#    "Operating System :: OS Independent",
  "Typing :: Typed",
]
dependencies = [
  "attrs >= 22.1.0",
  "cattrs >= 22.2.0",
  "gql[aiohttp,requests] >= 3.4.0",
  "strawberry-graphql >= 0.133.5",
  "typer[all] >= 0.6.1",
]

[project.urls]
"Homepage" = "https://dagger.io/"
"Bug Tracker" = "https://github.com/dagger/dagger/issues"

[project.scripts]
dagger-server-py = "dagger.server.cli:app"
dagger-py = "dagger.client.cli:app"

[tool.hatch.envs.default]
dependencies = [
  "pytest",
  "pytest-mock",
  "pytest-mypy",
  "pytest-sugar",
]

[tool.hatch.envs.default.scripts]
test = "pytest"
generate = "dagger-py generate --output src/dagger/client/api/gen.py"

[tool.hatch.envs.lint]
skip_install = true
dependencies = [
  "autoflake >=1.3.1",
  "black>=22.3.0",
  "flake8-black>=^0.3",
  "flake8-bugbear>=22.9.23",
  "flake8-eradicate>=^1.3.0",
  "flake8-isort>=^5.0.0",
  "flake8>=4.0.1",
  "isort>=5.10.1",
  "mypy>=0.942",
  "typing_extensions>=4.4.0",
]

[tool.hatch.envs.lint.scripts]
typing = "mypy --install-types --non-interactive src/dagger tests"
style = [
  "flake8 .",
  "black --check --diff .",
  "isort --check-only --diff .",
]
fmt = [
  "autoflake --in-place .",
  "isort .",
  "black .",
  "style",
]
all = [
  "style",
  "typing",
]

[tool.pytest.ini_options]
testpaths = ["tests/"]

[tool.mypy]
disallow_untyped_defs = false
follow_imports = "normal"
# ignore_missing_imports = true
warn_redundant_casts = true
pretty = true
show_column_numbers = true
warn_no_return = false
warn_unused_ignores = true
plugins = [
  "strawberry.ext.mypy_plugin",
]

[tool.black]
include = '\.pyi?$'
line-length = 119
target-version = ["py310", "py311"]

[tool.isort]
profile = "black"
line_length = 119
known_first_party = ["dagger"]

[tool.autoflake]
quiet = true
recursive = true
expand-star-imports = true
imports = ["graphql", "gql"]
remove-all-unused-imports = true
remove-duplicate-keys = true
remove-unused-variables = true
